+++ GHEtool/VariableClasses/Efficiency/__init__.py
@@ -2,3 +2,4 @@ from .COP import COP
 from .EER import EER
 from .SCOP import SCOP
 from .SEER import SEER
+from .EERCombined import EERCombined