+++ GHEtool/VariableClasses/LoadData/Baseclasses/_LoadDataBuilding.py
@@ -17,9 +17,10 @@ class _LoadDataBuilding(_LoadData, ABC):
 
     def __init__(self,
                  efficiency_heating: Union[int, float, COP, SCOP],
-                 efficiency_cooling: Union[int, float, EER, SEER],
+                 efficiency_cooling: Union[int, float, EER, SEER, EERCombined],
                  dhw: Union[float, np.ndarray] = None,
-                 efficiency_dhw: Union[int, float, COP, SCOP] = 4):
+                 efficiency_dhw: Union[int, float, COP, SCOP] = 4,
+                 multiyear: bool = False):
         """
 
         Parameters
@@ -32,8 +33,11 @@ class _LoadDataBuilding(_LoadData, ABC):
             Yearly value of array with energy demand for domestic hot water (DHW) [kWh]
         efficiency_dhw : int, float, COP, SCOP,
             Efficiency in DHW
+        multiyear : bool
+            True if multiyear data
         """
         super().__init__()
+        self._multiyear = multiyear
 
         # initiate variables
         self._baseload_heating: np.ndarray = np.zeros(12)
@@ -62,8 +66,7 @@ class _LoadDataBuilding(_LoadData, ABC):
         This function returns the monthly heating baseload in kWh/month for the whole simulation period.
 
         Returns
-        -------
-        baseload heating : np.ndarray
+        -------        baseload heating : np.ndarray
             Baseload heating for the whole simulation period
         """
 
@@ -244,7 +247,7 @@ class _LoadDataBuilding(_LoadData, ABC):
         return self._eer
 
     @eer.setter
-    def eer(self, efficiency_cooling: Union[int, float, EER, SEER]) -> None:
+    def eer(self, efficiency_cooling: Union[int, float, EER, SEER, EERCombined]) -> None:
         """
         This function defines the efficiency in cooling.
         Integer and float values will be automatically converted to an SEER object.
@@ -457,7 +460,7 @@ class _LoadDataBuilding(_LoadData, ABC):
         else:
             temperature = self.results.monthly_injection
 
-        return self.eer.get_EER(temperature, power=np.nan_to_num(power))
+        return self.eer.get_EER(temperature, power=np.nan_to_num(power), month_indices=self.month_indices)
 
     @staticmethod
     def conversion_factor_secondary_to_primary_heating(cop_value: Union[int, float, np.ndarray]) -> Union[
@@ -1010,3 +1013,14 @@ class _LoadDataBuilding(_LoadData, ABC):
         simulation period : int
         """
         return int(len(self.monthly_baseload_cooling_simulation_period) / 12)
+
+    @property
+    def month_indices(self) -> np.ndarray:
+        """
+        This property returns the array of all month indices for the simulation period.
+
+        Returns
+        -------
+        time array : np.ndarray
+        """
+        return np.tile(np.arange(1, 13), self.simulation_period)