+++ GHEtool/test/unit-tests/test_loads/test_hourly_load_building_multi_year.py
@@ -96,6 +96,7 @@ def test_yearly_loads_multiyear():
 
 
 def test_dhw():
+    HourlyBuildingLoadMultiYear(np.zeros(8760 * 10), np.ones(8760 * 10), 5, 20, np.ones(8760 * 10), 3)
     load = HourlyBuildingLoadMultiYear(np.zeros(8760 * 10), np.linspace(1, 8760 * 10 - 1, 8760 * 10) * 2, 6, 5)
 
     assert load.dhw == 0
@@ -145,3 +146,8 @@ def test_dhw():
     load.exclude_DHW_from_peak = True
     # idem since we started with an hourly data resolution
     assert np.allclose(load.monthly_peak_extraction_simulation_period, np.zeros(120))
+
+
+def test_time_array():
+    load = HourlyBuildingLoadMultiYear(np.zeros(8760 * 10), np.linspace(1, 8760 * 10 - 1, 8760 * 10) * 2, 6, 5)
+    assert np.allclose(load.month_indices, np.tile(np.repeat(np.arange(1, 13), load.UPM), 10))