+++ CHANGELOG.md
@@ -20,6 +20,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Efficiency classes for COP, EER, SCOP, SEER (issue #285).
 - Added building laod classes (issue #288).
 - Added __eq__ method for Result and Efficiency classes (issue #288).
+- Added _time_array to building loads (issue #291).
+- Added EERCombined for combined active and passive cooling efficiency (issue #291, #296).
+- Cluster Class (issue #298).
 
 ## Changed
 
@@ -39,6 +42,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   issue #283).
 - Removed a couple of log messages (issue #288).
 - Optimise load profile works with a variable COP/EER (issue #288).
+- Rename cylindrical_correction.py (issue #298).
 
 ## Fixed
 
@@ -47,6 +51,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Problem in CI/CD and testing for python <3.12 (issue #274).
 - Fix compatibility with numpy 2.0 (issue #274).
 - Fix problem with start month and zero peak loads (issue #288).
+- Problem with EERCombined (issue #300).
 
 ## [2.2.2] - 2024-05-16